remove comparative benchmarks from test
I see no reason to have this in the ongoing benchmarks moving forward in maintaining this library, and commenting it out aids in setting this library up for go module versioning without having an additional test dependency. They are preserved in comments if we need to run the comparative benchmarks again in the future.