chore: add go version to go.mod
As per the release notes of go 1.13, this seems to be required as a
feature gate for language changes. Running even `go test` on go1.13
tries to automatically add this in to the go.mod file, with the current
version of 1.13 being what it adds. However, we don't require any lang
features that came out since this code was written, so manually specify
1.10 instead.